refactor: move skill research capture logic
