+++ src/skills/research/signals.ts
