+++ src/skills/research/text.ts
